% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgn_adapter.R
\name{append_style_id}
\alias{append_style_id}
\title{Add an ID to an embedded render style.}
\usage{
append_style_id(document, style_id, item_id)
}
\arguments{
\item{document}{SBGN XML document.}

\item{style_id}{Style identifier.}

\item{item_id}{Glyph or arc identifier.}
}
\value{
Invisibly returns NULL.
}
\description{
Add an ID to an embedded render style.
}
