% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcose_layout.R
\name{build_adjacency}
\alias{build_adjacency}
\title{Build an adjacency list and connect disconnected components.}
\usage{
build_adjacency(node_count, indexed_edges)
}
\arguments{
\item{node_count}{Number of nodes.}

\item{indexed_edges}{Two-column integer matrix of zero-free R indexes.}
}
\value{
List of integer neighbor vectors.
}
\description{
Build an adjacency list and connect disconnected components.
}
