% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgn_adapter.R
\name{extract_top_level_glyphs}
\alias{extract_top_level_glyphs}
\title{Extract top-level glyph geometry.}
\usage{
extract_top_level_glyphs(map_node)
}
\arguments{
\item{map_node}{SBGN map node.}
}
\value{
Data frame of glyph geometry.
}
\description{
Extract top-level glyph geometry.
}
