% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgn_adapter.R
\name{format_coordinate}
\alias{format_coordinate}
\title{Format an SBGN coordinate.}
\usage{
format_coordinate(value)
}
\arguments{
\item{value}{Numeric coordinate.}
}
\value{
Compact character representation.
}
\description{
Format an SBGN coordinate.
}
