% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgn_adapter.R
\name{render_outputs}
\alias{render_outputs}
\title{Render an SBGN file with the installed native R renderer.}
\usage{
render_outputs(sbgn_path, png_path, svg_path)
}
\arguments{
\item{sbgn_path}{Input SBGN path.}

\item{png_path}{Output PNG path.}

\item{svg_path}{Output SVG path.}
}
\value{
NULL.
}
\description{
Render an SBGN file with the installed native R renderer.
}
