% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgn_adapter.R
\name{xml_number}
\alias{xml_number}
\title{Read one numeric XML attribute.}
\usage{
xml_number(node, name)
}
\arguments{
\item{node}{xml2 node.}

\item{name}{Attribute name.}
}
\value{
Numeric attribute value.
}
\description{
Read one numeric XML attribute.
}
